fix: set item.scrollIntoView block to nearest
